Extract create method generator to separate class
