+++ kotlin-inject-compiler-core/src/main/kotlin/me/tatarka/inject/compiler/InjectGenerator.kt
@@ -5,22 +5,21 @@ import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.plusParameter
 import com.squareup.kotlinpoet.PropertySpec
-import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
-import com.squareup.kotlinpoet.asClassName
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Inject
 import me.tatarka.inject.annotations.Provides
 import me.tatarka.inject.annotations.Scope
-import kotlin.reflect.KClass
 
 private val SCOPED_COMPONENT = ClassName("me.tatarka.inject.internal", "ScopedComponent")
 private val LAZY_MAP = ClassName("me.tatarka.inject.internal", "LazyMap")
 
-class InjectGenerator(provider: AstProvider, private val options: Options) :
-    AstProvider by provider {
+class InjectGenerator(
+    provider: AstProvider,
+    private val options: Options,
+    private val createGenerator: CreateGenerator
+) : AstProvider by provider {
 
     var scopeType: AstType? = null
         private set
@@ -35,7 +34,7 @@ class InjectGenerator(provider: AstProvider, private val options: Options) :
         val constructor = astClass.primaryConstructor
 
         val injectComponent = generateInjectComponent(astClass, constructor)
-        val createFunction = generateCreate(astClass, constructor, injectComponent)
+        val createFunction = createGenerator.create(astClass, constructor, injectComponent)
 
         return FileSpec.builder(astClass.packageName, "Inject${astClass.name}")
             .addType(injectComponent)
@@ -150,93 +149,6 @@ class InjectGenerator(provider: AstProvider, private val options: Options) :
                 }
             }
             .build()
-    }
-
-    @Suppress("ComplexMethod")
-    private fun generateCreate(
-        element: AstClass,
-        constructor: AstConstructor?,
-        injectComponent: TypeSpec
-    ): List<FunSpec> {
-        fun generateCreate(
-            typeName: TypeName,
-            constructor: AstConstructor?,
-            injectComponent: TypeSpec,
-            companion: AstClass?,
-            params: List<AstParam>
-        ): FunSpec {
-            return FunSpec.builder("create")
-                .apply {
-                    if (constructor != null) {
-                        for (param in params) {
-                            addParameter(param.asParameterSpec())
-                        }
-                    }
-                    if (companion != null) {
-                        receiver(companion.type.asTypeName())
-                    } else {
-                        receiver(KClass::class.asClassName().plusParameter(typeName))
-                    }
-                }
-                .returns(typeName)
-                .apply {
-                    val codeBlock = CodeBlock.builder()
-                    codeBlock.add("return %N(", injectComponent)
-                    if (constructor != null) {
-                        params.forEachIndexed { i, parameter ->
-                            if (i != 0) {
-                                codeBlock.add(", ")
-                            }
-                            codeBlock.add("%L", parameter.name)
-                        }
-                    }
-                    codeBlock.add(")")
-                    addCode(codeBlock.build())
-                }
-                .build()
-        }
-
-        @Suppress("MaxLineLength")
-        val companion = if (options.generateCompanionExtensions) {
-            element.companion.also {
-                if (it == null) {
-                    error(
-                        """Missing companion for class: ${element.asClassName()}.
-                            |When you have the option me.tatarka.inject.generateCompanionExtensions=true you must declare a companion option on the component class for the extension function to apply to.
-                            |You can do so by adding 'companion object' to the class.
-                        """.trimMargin(),
-                        element
-                    )
-                }
-            }
-        } else {
-            null
-        }
-        return mutableListOf<FunSpec>().apply {
-            val typeName = element.type.asTypeName()
-            val params = constructor?.parameters ?: emptyList()
-            add(
-                generateCreate(
-                    typeName,
-                    constructor,
-                    injectComponent,
-                    companion,
-                    params
-                )
-            )
-            val nonDefaultParams = constructor?.parameters?.filter { !it.hasDefault } ?: emptyList()
-            if (params.size != nonDefaultParams.size) {
-                add(
-                    generateCreate(
-                        typeName,
-                        constructor,
-                        injectComponent,
-                        companion,
-                        nonDefaultParams
-                    )
-                )
-            }
-        }
     }
 
     private fun collectTypes(
